[Issue: 103] Refactor mode-specific traitlets   Move mode-specific traitlets into their respective personalities.
  Personality traitlets created with the tag "config=True" will have
  their values set into the web app's settings dictionary, keeping
  handlers decoupled from any specific personalities.
(c) Copyright IBM Corp. 2016
